Moved AuthorizedRegistryScenarion and test back
